requirements.txt
appdirs==1.4.0
boto3==1.4.4
botocore==1.5.24
click==6.7
docutils==0.13.1
Flask==0.12
futures==3.0.5
itsdangerous==0.24
Jinja2==2.9.5
jmespath==0.9.2
MarkupSafe==0.23
oauthlib==2.0.1
packaging==16.8
pyparsing==2.1.10
python-dateutil==2.6.0
requests==2.13.0
requests-oauthlib==0.8.0
s3transfer==0.1.10
six==1.10.0
tweepy==3.5.0
Werkzeug==0.11.15
